added basic save/load for segments

+++ src/main/java/fr/alexandredch/vectours/data/Vector.java
@@ -2,4 +2,4 @@ package fr.alexandredch.vectours.data;
 
 import java.io.Serializable;
 
-public record Vector(String id, float[] values, Metadata metadata) implements Serializable {}
+public record Vector(String id, double[] values, Metadata metadata) implements Serializable {}
