+++ src/main/java/fr/alexandredch/vectours/serialization/InMemorySerializer.java
@@ -4,13 +4,13 @@ import java.io.*;
 
 public final class InMemorySerializer {
 
-    public static float[] deserialize(byte[] bytes) throws IOException, ClassNotFoundException {
+    public static double[] deserialize(byte[] bytes) throws IOException, ClassNotFoundException {
         ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(bytes);
         ObjectInputStream objectInputStream = new ObjectInputStream(byteArrayInputStream);
-        return (float[]) objectInputStream.readObject();
+        return (double[]) objectInputStream.readObject();
     }
 
-    public static byte[] serialize(float[] vector) throws IOException {
+    public static byte[] serialize(double[] vector) throws IOException {
         ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
         ObjectOutputStream objectOutputStream = new ObjectOutputStream(byteArrayOutputStream);
         objectOutputStream.writeObject(vector);
